refactor: clean up unit test directory
