+++ test/unit/assorted/polling_srv_records_for_mongos_discovery.prose.test.ts
@@ -2,11 +2,11 @@ import { expect } from 'chai';
 import * as dns from 'dns';
 import * as sinon from 'sinon';
 
-import { MongoClient } from '../../src';
-import { isHello } from '../../src/utils';
-import * as mock from '../tools/mongodb-mock/index';
-import type { MockServer } from '../tools/mongodb-mock/src/server';
-import { processTick } from '../tools/utils';
+import { MongoClient } from '../../../src';
+import { isHello } from '../../../src/utils';
+import * as mock from '../../tools/mongodb-mock/index';
+import type { MockServer } from '../../tools/mongodb-mock/src/server';
+import { processTick } from '../../tools/utils';
 
 /*
     The SRV Prose Tests make use of the following REAL DNS records.
